# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "BestFriend API",
    "description": "",
    "version": "1.0.0"
  },
  "host": "mishka-l3tq.onrender.com",
  "basePath": "",
  "schemes": [
    "https"
  ],
  "paths": {
    "/generate-code": {
      "post": {
        "tags": ["Регистрация"],
        "description": "Генерация кода для регистрации",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/verify-code": {
      "post": {
        "tags": ["Регистрация"],
        "description": "Проверка кода для регистрации",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "code": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/upload": {
      "post": {
        "tags": ["Загрузка"],
        "description": "Загрузка файла на сервер для сохранения",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/create": {
      "post": {
        "tags": ["Готовые аудио"],
        "description": "Создание записи готового аудио",
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/all": {
      "get": {
        "tags": ["Готовые аудио"],
        "description": "Получение всех готовых аудио",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/delete/{id}": {
      "delete": {
        "tags": ["Записи пользователя"],
        "description": "Удаление записи пользователя",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/create_record": {
      "post": {
        "tags": ["Записи пользователя"],
        "description": "После upload надо создать запись пользователя",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "audioFile": {
                  "example": "any"
                },
                "duration": {
                  "example": "any"
                },
                "user": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user-records": {
      "post": {
        "tags": ["Записи пользователя"],
        "description": "Получение всех аудио пользователя",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}